+++ packages/nikita/lib/index.js
@@ -1,2 +1,8 @@
 
+require('@nikita/db/lib/register')
+require('@nikita/docker/lib/register')
+require('@nikita/filetypes/lib/register')
+require('@nikita/krb5/lib/register')
+require('@nikita/ldap/lib/register')
+require('@nikita/tools/lib/register')
 module.exports = require('@nikita/core')